requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9
pydantic-settings>=2.0
python-dotenv>=1.0
anthropic>=0.40.0
google-genai>=1.16.0
google-cloud-firestore>=2.16.0
alembic>=1.13.0
websockets>=12.0
httpx>=0.27.0
sendgrid>=6.11.0
